// src/generated/models/Cr38e_masterorderingcatalogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr38e_masterorderingcatalogscr38e_department = {
  357090000: 'IT',
  357090001: 'Finance',
  357090002: 'Operations',
  357090003: 'HR',
  357090004: 'Engineering',
  357090005: 'Marketing'
} as const;
export type Cr38e_masterorderingcatalogscr38e_department = keyof typeof Cr38e_masterorderingcatalogscr38e_department;
export const Cr38e_masterorderingcatalogscr38e_um = {
  357090000: 'Each',
  357090001: 'Box',
  357090002: 'Case',
  357090003: 'Dozen',
  357090004: 'Pound',
  357090005: 'Kilogram',
  357090006: 'Liter',
  357090007: 'Gallon'
} as const;
export type Cr38e_masterorderingcatalogscr38e_um = keyof typeof Cr38e_masterorderingcatalogscr38e_um;
export const Cr38e_masterorderingcatalogsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr38e_masterorderingcatalogsstatecode = keyof typeof Cr38e_masterorderingcatalogsstatecode;
export const Cr38e_masterorderingcatalogsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr38e_masterorderingcatalogsstatuscode = keyof typeof Cr38e_masterorderingcatalogsstatuscode;

export interface Cr38e_masterorderingcatalogsBase {
  cr38e_catalognumber: string;
  cr38e_category: string;
  cr38e_department?: Cr38e_masterorderingcatalogscr38e_department;
  cr38e_description: string;
  cr38e_imageurl?: string;
  cr38e_keywords?: string;
  cr38e_masterorderingcatalogid: string;
  cr38e_partnumber?: string;
  cr38e_sortcode?: string;
  cr38e_um?: Cr38e_masterorderingcatalogscr38e_um;
  cr38e_vendoraltlink?: string;
  cr38e_vendorlink?: string;
  cr38e_vendorname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr38e_masterorderingcatalogsstatecode;
  statuscode?: Cr38e_masterorderingcatalogsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr38e_masterorderingcatalogs extends Cr38e_masterorderingcatalogsBase {
  cr38e_departmentname?: string;
  cr38e_umname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}